# frontend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ILMA API
  version: 1.0.0
  description: |
    API for ILMA Education Platform (Offline-First).
    Targets CM2 students in Benin preparing for CEP exam.
    Three roles: Student, Parent, Admin.
servers:
  - url: https://api.ilma.app/v1
    description: Production
  - url: https://staging-api.ilma.app/v1
    description: Staging
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Domain:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string }
        name: { type: string }
        description: { type: string }
        progress: { type: number }
        skillCount: { type: integer }
    Skill:
      type: object
      properties:
        id: { type: string }
        domainId: { type: string }
        name: { type: string }
        smartScore: { type: number, minimum: 0, maximum: 100 }
        status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, MASTERED] }
        exerciseCount: { type: integer }
...
security:
  - bearerAuth: []
paths:
...
  /subjects/{subjectId}/domains:
    get:
      tags: [Content]
      summary: List domains for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Domains list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Domain' }
  /domains/{domainId}/skills:
    get:
      tags: [Content]
      summary: List skills for a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Skill' }
...